requirements.txt
fastapi==0.103.2
uvicorn==0.22.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
Pillow==10.0.1
torch==2.0.1
transformers==4.33.2